+++ src-gui/src/renderer/rpc.ts
@@ -89,14 +89,22 @@ export async function withdrawBtc(address: string): Promise<string> {
 
 export async function buyXmr(
   seller: Provider,
-  bitcoin_change_address: string,
+  bitcoin_change_address: string | null,
   monero_receive_address: string,
 ) {
-  await invoke<BuyXmrArgs, BuyXmrResponse>("buy_xmr", {
-    seller: providerToConcatenatedMultiAddr(seller),
-    bitcoin_change_address,
-    monero_receive_address,
-  });
+  await invoke<BuyXmrArgs, BuyXmrResponse>(
+    "buy_xmr",
+    bitcoin_change_address == null
+      ? {
+          seller: providerToConcatenatedMultiAddr(seller),
+          monero_receive_address,
+        }
+      : {
+          seller: providerToConcatenatedMultiAddr(seller),
+          monero_receive_address,
+          bitcoin_change_address,
+        },
+  );
 }
 
 export async function resumeSwap(swapId: string) {