+++ swap/src/cli/command.rs
@@ -78,8 +78,10 @@ where
         } => {
             let monero_receive_address =
                 monero_address::validate_is_testnet(monero_receive_address, is_testnet)?;
-            let bitcoin_change_address =
-                bitcoin_address::validate_is_testnet(bitcoin_change_address, is_testnet)?;
+
+            let bitcoin_change_address = bitcoin_change_address
+                .map(|address| bitcoin_address::validate_is_testnet(address, is_testnet))
+                .transpose()?;
 
             let context = Arc::new(
                 ContextBuilder::new(is_testnet)
@@ -372,10 +374,10 @@ enum CliCommand {
 
         #[structopt(
             long = "change-address",
-            help = "The bitcoin address where any form of change or excess funds should be sent to",
+            help = "The bitcoin address where any form of change or excess funds should be sent to. If omitted they will be sent to the internal wallet.",
             parse(try_from_str = bitcoin_address::parse)
         )]
-        bitcoin_change_address: bitcoin::Address,
+        bitcoin_change_address: Option<bitcoin::Address>,
 
         #[structopt(flatten)]
         monero: Monero,